+++ engelsystem-base/src/test/kotlin/info/metadude/kotlin/library/engelsystem/ProductionApiTest.kt
@@ -37,7 +37,9 @@ class ProductionApiTest {
 
     private fun assertShift(shift: Shift) {
         assertThat(shift.sID).isGreaterThan(0)
+        assertThat(shift.typeDescription).isNotNull()
         assertThat(shift.typeId).isGreaterThan(0)
+        assertThat(shift.typeName).isNotEmpty()
         assertThat(shift.name).isNotEmpty()
         assertThat(shift.locationName).isNotEmpty()
         assertThat(shift.locationDescription).isNotNull()